+++ SecretSanta/src/SecretSanta.Domain/Models/Message.cs
@@ -1,0 +1,17 @@
+using System.Collections.Generic;
+
+namespace SecretSanta.Domain
+{
+    public class Message
+    {
+        string Content { set; get; }
+
+        public Message(string message)
+        {
+            Content = message;
+        }
+
+
+
+    }
+}
