+++ SecretSanta/src/SecretSanta.Domain/Models/Pairing.cs
@@ -1,0 +1,16 @@
+namespace SecretSanta.Domain
+{
+    public class Pairing
+    {
+        User Recipient { set; get; }
+        User Santa { set; get; }
+
+        public Pairing(User recipient, User santa)
+        {
+            Recipient = recipient;
+            Santa = santa;
+        }
+
+
+    }
+}
