+++ SecretSanta/src/SecretSanta.Domain/Services/GiftServices.cs
@@ -1,0 +1,20 @@
+namespace SecretSanta.Domain
+{
+    public class GiftService
+    {
+        private void CreateGift(User theUser)
+        {
+
+        }
+
+        private void EditGift(User theUser)
+        {
+
+        }
+
+        private void DeleteGift(User theUser)
+        {
+
+        }
+    }
+}
