+++ SecretSanta/src/SecretSanta.Domain/Services/GroupServices.cs
@@ -1,0 +1,20 @@
+namespace SecretSanta.Domain
+{
+    public class GroupService
+    {
+        private void CreateGroup()
+        {
+
+        }
+
+        private void AddUser()
+        {
+
+        }
+
+        private void RemoveUser()
+        {
+
+        }
+    }
+}
