+++ SecretSanta/src/SecretSanta.Domain/Services/MessageServices.cs
@@ -1,0 +1,10 @@
+namespace SecretSanta.Domain
+{
+    public class MessageService
+    {
+        public void AddMessage(string message, string id)
+        {
+            //pin id to start of each message - id either "Santa" or "Recipient"
+        }
+    }
+}
