+++ SecretSanta/src/SecretSanta.Domain/Services/UserServices.cs
@@ -1,0 +1,15 @@
+namespace SecretSanta.Domain
+{
+    public class UserServices
+    {
+        public void CreateUser()
+        {
+
+        }
+
+        public void UpdateUser()
+        {
+
+        }
+    }
+}
